Adopt README.md so it reflects current status for master

+++ ctp2_code/libs/anet/src/dp/dp2.c
@@ -253,6 +253,7 @@ MSVC's warning level is set to 4.
 #include <sys/types.h>
 #include <unistd.h>
 #endif
+#include <stdint.h>
 
 #ifdef _WIN32
 #include <crtdbg.h>
@@ -1103,7 +1104,7 @@ DP_API int dp_unpack_session(dp_t *dp, const char *subkey, int subkeylen, const 
 	if (len > 0)
 		memcpy(p->szUserField, userbuf+4, len);
 
-	len = buflen - ((int)q - (int)buf);
+	len = buflen - ((uintptr_t)q - (uintptr_t)buf);
 	if (len > 0) {
 		/* session has migrated.  Receive new master address. */
 		assert(len >= dp->dpio->myAdrLen);
@@ -1116,7 +1117,7 @@ DP_API int dp_unpack_session(dp_t *dp, const char *subkey, int subkeylen, const 
 
 	dprint_sess(dp, p, "dp_unpack_session (new)");
 	dp_assertValid(dp);
-	return ((int)q - (int)buf);
+	return ((uintptr_t)q - (uintptr_t)buf);
 }
 #else
 
@@ -1307,7 +1308,7 @@ static int dp_unpack_host(dpid_t firstId, int myAdrLen, const char *buf, dp_host
 	printAdr(myAdrLen, p->iadr);
 
 	/* Depending upon the size of this host a second address could be embeded */
-	if (length > ((int)q - (int)buf)) {
+	if (length > ((uintptr_t)q - (uintptr_t)buf)) {
 		memcpy(p->iadr2, q, myAdrLen);
 		q += myAdrLen;
 
@@ -1320,7 +1321,7 @@ static int dp_unpack_host(dpid_t firstId, int myAdrLen, const char *buf, dp_host
 	p->firstId = firstId;
 	DPRINT(("\n"));
 
-	return ((int)q - (int)buf);
+	return ((uintptr_t)q - (uintptr_t)buf);
 }
 
 /*----------------------------------------------------------------------
@@ -1350,7 +1351,7 @@ static int dp_pack_host(int myAdrLen, const dp_host_t *p, char *buf)
 		printAdr(myAdrLen, p->iadr2);
 	}
 	DPRINT(("\n"));
-	return ((int)q - (int)buf);
+	return ((uintptr_t)q - (uintptr_t)buf);
 }
 
 /*----------------------------------------------------------------------
@@ -1433,7 +1434,7 @@ int dp_unpack_playerId(dpid_t id, const char *buf, dp_playerId_t *p)
 	DPRINT(("dp_unpack_playerId(id %d): name %s\n", id, p->name));
 #endif
 
-  return ((int)q - (int)buf);
+  return ((uintptr_t)q - (uintptr_t)buf);
 }
 
 /*--------------------------------------------------------------------------
@@ -3410,7 +3411,7 @@ DP_API dp_result_t dpCommThaw(dp_t **pdp, FILE *thawfp, dpCommThawCallback_t cb,
 	if (params.OpenAddresses[0]) {
 		DPRINT(("dpCommThaw: Open Addresses %s\n", params.OpenAddresses));
 		params.commInitReq.flags |= comm_INIT_FLAGS_CONN_ADDR;
-		params.commInitReq.dialing_method = (int) params.OpenAddresses;
+		params.commInitReq.dialing_method = (uintptr_t) params.OpenAddresses;
 	}
 
 	/* Load Activenet. */
@@ -4014,7 +4015,7 @@ DP_API dp_result_t dpResolveHostname(
 		int i;
 		dp_serverInfo_t *server;
 		char subkey[dptab_KEY_MAXLEN];
-		int serverlen;
+		size_t serverlen;
 		int subkeylen;
 		for (i = 0; i < dptab_tableSize(dp->serverpings); i++) {
 			if (dptab_get_byindex(dp->serverpings, i, (void **)&server,
@@ -5188,7 +5189,7 @@ dpHandleJoinSession(
 		char subkey[dp_KEY_MAXLEN];
 		int subkeylen = 2;
 		char *buf;
-		int buflen;
+		size_t buflen;
 
 		host.firstId = (dpid_t) (dp_PLAYERS_PER_HOST * sp->hostid++);
 		/* Wrap dpids around at dp_MAXDPIDS back to dp_FIRST_DPID
@@ -8234,7 +8235,7 @@ static dp_result_t dp_election_become_master(dp_t *dp, dpid_t winner_id)
 	 */
 	sp->hostid = (highest_id / dp_PLAYERS_PER_HOST) + 2;
 	DPRINT(("dp_election_become_master: new players will start at id:%d\n", sp->hostid * dp_PLAYERS_PER_HOST));
-	{	int i; int len; int subkeylen;
+	{	int i; size_t len; size_t subkeylen;
 		char subkey[dptab_KEY_MAXLEN];
 		char *buf;
 		subkeylen = 2;
@@ -9164,7 +9165,7 @@ dp_result_t dp_uid2sessid(dp_t *dp, dp_uid_t uid, char *sessidbuf, int *sessidle
 	char subkey[dptab_KEY_MAXLEN];
 	char *pbuf;
 	dp_result_t err;
-	int len;
+	size_t len;
 
 	subkey[0] = dpGETLONG_FIRSTBYTE(uid);
 	subkey[1] = dpGETLONG_SECONDBYTE(uid);
@@ -9646,7 +9647,7 @@ static dp_result_t dp_receive(
 			char subkey[dptab_KEY_MAXLEN];
 			int subkeylen = 0;
 			char *buf;
-			int buflen;
+			size_t buflen;
 			int idoffset = (dpid_t)pkt->body.pData.id - firstId;
 			subkey[subkeylen++] = (char) dpGETSHORT_FIRSTBYTE(idoffset);
 			subkey[subkeylen++] = (char) dpGETSHORT_FIRSTBYTE(pkt->body.pData.key);
